+++ Assets/Scripts/Weapon.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Weapon : MonoBehaviour
 {
-   protected int quantityDamage = 1;
+   protected float quantityDamage = 1f;
    protected float range = 4.0f;
    protected float cooldown;
    protected bool onCooldown = false;
